\name{qtables}
\docType{data}
\alias{q.testlin.n2}
\alias{q.testlin.n3}
\alias{q.testlin.n4}
\alias{q.testlin.n5}
\alias{q.testlin.n6}
\alias{q.testlin.n7}
\alias{q.testlin.n8}
\alias{q.testlin.n9}
\alias{q.testlin.n10}
\alias{q.testlin.n11}
\alias{q.testlin.n12}
\alias{q.testlin.n13}
\alias{q.testlin.n14}
\alias{q.testlin.n15}
\alias{q.testlin.n16}
\alias{q.testlin.n17}
\alias{q.testlin.n18}
\alias{q.testlin.n19}
\alias{q.testlin.n20}
\alias{q.testlin.n21}
\alias{q.testlin.n22}
\alias{q.testlin.n23}
\alias{q.testlin.n24}
\alias{q.testlin.n25}
\alias{q.testlin.n26}
\alias{q.testlin.n27}
\alias{q.testlin.n28}
\alias{q.testlin.n29}
\alias{q.testlin.n30}
\alias{q.testlin.n31}
\alias{q.testlin.n32}
\alias{q.testlin.n33}
\alias{q.testlin.n34}
\alias{q.testlin.n35}
\alias{q.testlin.n36}
\alias{q.testlin.n37}
\alias{q.testlin.n38}
\alias{q.testlin.n39}
\alias{q.testlin.n40}
\alias{q.testlin.n41}
\alias{q.testlin.n42}
\alias{q.testlin.n43}
\alias{q.testlin.n44}
\alias{q.testlin.n45}
\alias{q.testlin.n46}
\alias{q.testlin.n47}
\alias{q.testlin.n48}
\alias{q.testlin.n49}
\alias{q.testlin.n50}
\alias{q.testgaplin.n2}
\alias{q.testgaplin.n3}
\alias{q.testgaplin.n4}
\alias{q.testgaplin.n5}
\alias{q.testgaplin.n6}
\alias{q.testgaplin.n7}
\alias{q.testgaplin.n8}
\alias{q.testgaplin.n9}
\alias{q.testgaplin.n10}
\alias{q.testgaplin.n11}
\alias{q.testgaplin.n12}
\alias{q.testgaplin.n13}
\alias{q.testgaplin.n14}
\alias{q.testgaplin.n15}
\alias{q.testgaplin.n16}
\alias{q.testgaplin.n17}
\alias{q.testgaplin.n18}
\alias{q.testgaplin.n19}
\alias{q.testgaplin.n20}
\alias{q.testgaplin.n21}
\alias{q.testgaplin.n22}
\alias{q.testgaplin.n23}
\alias{q.testgaplin.n24}
\alias{q.testgaplin.n25}
\alias{q.testgaplin.n26}
\alias{q.testgaplin.n27}
\alias{q.testgaplin.n28}
\alias{q.testgaplin.n29}
\alias{q.testgaplin.n30}
\alias{q.testgaplin.n31}
\alias{q.testgaplin.n32}
\alias{q.testgaplin.n33}
\alias{q.testgaplin.n34}
\alias{q.testgaplin.n35}
\alias{q.testgaplin.n36}
\alias{q.testgaplin.n37}
\alias{q.testgaplin.n38}
\alias{q.testgaplin.n39}
\alias{q.testgaplin.n40}
\alias{q.testgaplin.n41}
\alias{q.testgaplin.n42}
\alias{q.testgaplin.n43}
\alias{q.testgaplin.n44}
\alias{q.testgaplin.n45}
\alias{q.testgaplin.n46}
\alias{q.testgaplin.n47}
\alias{q.testgaplin.n48}
\alias{q.testgaplin.n49}
\alias{q.testgaplin.n50}
\title{Quantile Tables of the Linearity/Gap-Linearity Tests}
\description{
  The data set provide quantile tables for the linearity/gap-linearity test
statistics for N=2, ..., 50 and k= 2, ..., N, for each N. These tables will be
used to compute p-values corresponding to test statistics.
}
\usage{qtables}
\format{R rda files. Within each quantile table, the first row indicates at
which probability values the quantiles are computed.}
\keyword{datasets}
